[phrase match] Restructure `full_text_index::inverted_index` module (#6638)
* move all inverted indexes to one module

Also, remove `old_mmap_postings` and `compressed_postings`

* move congruence test into its own file

* remove old de/serialize_tokenset functions
