+++ lib/segment/src/index/field_index/full_text_index/immutable_text_index.rs
@@ -2,8 +2,10 @@ use std::path::PathBuf;
 
 use common::types::PointOffsetType;
 
-use super::immutable_inverted_index::ImmutableInvertedIndex;
+use super::inverted_index::immutable_inverted_index::ImmutableInvertedIndex;
 use super::inverted_index::InvertedIndex;
+#[cfg(feature = "rocksdb")]
+use super::inverted_index::mutable_inverted_index::MutableInvertedIndex;
 use super::mmap_text_index::MmapFullTextIndex;
 #[cfg(feature = "rocksdb")]
 use super::text_index::FullTextIndex;
@@ -12,7 +14,7 @@ use crate::common::operation_error::{OperationError, OperationResult};
 #[cfg(feature = "rocksdb")]
 use crate::common::rocksdb_buffered_delete_wrapper::DatabaseColumnScheduledDeleteWrapper;
 use crate::data_types::index::TextIndexParams;
-use crate::index::field_index::full_text_index::mmap_inverted_index::mmap_postings_enum::MmapPostingsEnum;
+use crate::index::field_index::full_text_index::inverted_index::mmap_inverted_index::mmap_postings_enum::MmapPostingsEnum;
 
 pub struct ImmutableFullTextIndex {
     pub(super) inverted_index: ImmutableInvertedIndex,
@@ -92,10 +94,7 @@ impl ImmutableFullTextIndex {
             Ok((idx, tokens))
         });
 
-        let mutable = super::mutable_inverted_index::MutableInvertedIndex::build_index(
-            iter,
-            phrase_matching,
-        )?;
+        let mutable = MutableInvertedIndex::build_index(iter, phrase_matching)?;
 
         self.inverted_index = ImmutableInvertedIndex::from(mutable);
         Ok(true)
