+++ lib/segment/src/index/field_index/full_text_index/inverted_index/mod.rs
@@ -1,3 +1,12 @@
+pub(super) mod immutable_inverted_index;
+pub mod immutable_postings_enum;
+pub(super) mod mmap_inverted_index;
+pub(super) mod mutable_inverted_index;
+pub(super) mod mutable_inverted_index_builder;
+mod positions;
+mod posting_list;
+mod postings_iterator;
+
 use std::collections::HashMap;
 
 use ahash::AHashSet;
@@ -286,9 +295,9 @@ mod tests {
     use rstest::rstest;
 
     use super::{Document, InvertedIndex, ParsedQuery, TokenId, TokenSet};
-    use crate::index::field_index::full_text_index::immutable_inverted_index::ImmutableInvertedIndex;
-    use crate::index::field_index::full_text_index::mmap_inverted_index::MmapInvertedIndex;
-    use crate::index::field_index::full_text_index::mutable_inverted_index::MutableInvertedIndex;
+    use crate::index::field_index::full_text_index::inverted_index::immutable_inverted_index::ImmutableInvertedIndex;
+    use crate::index::field_index::full_text_index::inverted_index::mmap_inverted_index::MmapInvertedIndex;
+    use crate::index::field_index::full_text_index::inverted_index::mutable_inverted_index::MutableInvertedIndex;
 
     fn generate_word() -> String {
         let mut rng = rand::rng();
