+++ lib/segment/src/index/field_index/full_text_index/inverted_index/mutable_inverted_index_builder.rs
@@ -1,6 +1,6 @@
 use common::types::PointOffsetType;
 
-use super::inverted_index::InvertedIndex;
+use super::InvertedIndex;
 use super::mutable_inverted_index::MutableInvertedIndex;
 #[cfg(feature = "rocksdb")]
 use crate::common::operation_error::OperationResult;
