+++ lib/segment/src/index/field_index/full_text_index/mod.rs
@@ -1,19 +1,9 @@
-mod immutable_postings_enum;
 mod immutable_text_index;
 mod inverted_index;
-mod mmap_inverted_index;
 pub mod mmap_text_index;
 mod mutable_text_index;
-mod positions;
-mod posting_list;
-mod postings_iterator;
 pub mod text_index;
 mod tokenizers;
 
-#[cfg(test)]
-mod compressed_posting;
-mod immutable_inverted_index;
-mod mutable_inverted_index;
-mod mutable_inverted_index_builder;
 #[cfg(test)]
 mod tests;
