+++ lib/segment/src/index/field_index/full_text_index/tests/test_congruence.rs
@@ -1,0 +1,392 @@
+use std::collections::HashSet;
+use std::sync::Arc;
+
+use common::counter::hardware_counter::HardwareCounterCell;
+use common::types::PointOffsetType;
+use parking_lot::RwLock;
+use rand::SeedableRng;
+use rand::rngs::StdRng;
+use rocksdb::DB;
+use rstest::rstest;
+use serde_json::Value;
+use tempfile::{Builder, TempDir};
+
+use crate::common::operation_error::OperationResult;
+use crate::common::rocksdb_wrapper::open_db_with_existing_cf;
+use crate::data_types::index::TextIndexParams;
+use crate::fixtures::payload_fixtures::random_full_text_payload;
+use crate::index::field_index::field_index_base::PayloadFieldIndex;
+use crate::index::field_index::full_text_index::immutable_text_index::ImmutableFullTextIndex;
+use crate::index::field_index::full_text_index::inverted_index::{
+    InvertedIndex, ParsedQuery, TokenId, TokenSet,
+};
+use crate::index::field_index::full_text_index::mmap_text_index::FullTextMmapIndexBuilder;
+use crate::index::field_index::full_text_index::mutable_text_index::{self, MutableFullTextIndex};
+use crate::index::field_index::full_text_index::text_index::{
+    FullTextGridstoreIndexBuilder, FullTextIndex, FullTextIndexBuilder,
+};
+use crate::index::field_index::{FieldIndexBuilderTrait, ValueIndexer};
+use crate::json_path::JsonPath;
+
+const FIELD_NAME: &str = "test";
+const TYPES: &[IndexType] = &[
+    #[cfg(feature = "rocksdb")]
+    IndexType::Mutable,
+    IndexType::MutableGridstore,
+    #[cfg(feature = "rocksdb")]
+    IndexType::Immutable,
+    IndexType::Mmap,
+    IndexType::RamMmap,
+];
+
+#[derive(Clone, Copy, PartialEq, Debug)]
+enum IndexType {
+    #[cfg(feature = "rocksdb")]
+    Mutable,
+    MutableGridstore,
+    #[cfg(feature = "rocksdb")]
+    Immutable,
+    Mmap,
+    RamMmap,
+}
+
+enum IndexBuilder {
+    #[cfg(feature = "rocksdb")]
+    Mutable(FullTextIndexBuilder),
+    MutableGridstore(FullTextGridstoreIndexBuilder),
+    #[cfg(feature = "rocksdb")]
+    Immutable(FullTextIndexBuilder),
+    Mmap(FullTextMmapIndexBuilder),
+    RamMmap(FullTextMmapIndexBuilder),
+}
+
+impl IndexBuilder {
+    fn add_point(
+        &mut self,
+        id: PointOffsetType,
+        payload: &[&Value],
+        hw_counter: &HardwareCounterCell,
+    ) -> OperationResult<()> {
+        match self {
+            #[cfg(feature = "rocksdb")]
+            IndexBuilder::Mutable(builder) => builder.add_point(id, payload, hw_counter),
+            IndexBuilder::MutableGridstore(builder) => {
+                FieldIndexBuilderTrait::add_point(builder, id, payload, hw_counter)
+            }
+            #[cfg(feature = "rocksdb")]
+            IndexBuilder::Immutable(builder) => builder.add_point(id, payload, hw_counter),
+            IndexBuilder::Mmap(builder) => {
+                FieldIndexBuilderTrait::add_point(builder, id, payload, hw_counter)
+            }
+            IndexBuilder::RamMmap(builder) => {
+                FieldIndexBuilderTrait::add_point(builder, id, payload, hw_counter)
+            }
+        }
+    }
+
+    fn finalize(self) -> OperationResult<FullTextIndex> {
+        match self {
+            #[cfg(feature = "rocksdb")]
+            IndexBuilder::Mutable(builder) => builder.finalize(),
+            IndexBuilder::MutableGridstore(builder) => builder.finalize(),
+            #[cfg(feature = "rocksdb")]
+            IndexBuilder::Immutable(builder) => {
+                let FullTextIndex::Mutable(index) = builder.finalize()? else {
+                    panic!("expected mutable index");
+                };
+
+                // Deconstruct mutable index, flush pending changes
+                let MutableFullTextIndex {
+                    storage,
+                    inverted_index: _,
+                    config,
+                } = index;
+                let mutable_text_index::Storage::RocksDb(db_wrapper) = storage else {
+                    panic!("expected RocksDB storage for immutable index");
+                };
+                db_wrapper.flusher()().expect("failed to flush");
+
+                // Open and load immutable index
+                let mut index = ImmutableFullTextIndex::open_rocksdb(db_wrapper, config);
+                index.load()?;
+                let index = FullTextIndex::Immutable(index);
+                Ok(index)
+            }
+            IndexBuilder::Mmap(builder) => builder.finalize(),
+            IndexBuilder::RamMmap(builder) => {
+                let FullTextIndex::Mmap(index) = builder.finalize()? else {
+                    panic!("expected mmap index");
+                };
+
+                // Load index from mmap
+                let mut index = FullTextIndex::Immutable(ImmutableFullTextIndex::open_mmap(*index));
+                index.load()?;
+                Ok(index)
+            }
+        }
+    }
+}
+
+fn create_builder(
+    index_type: IndexType,
+    phrase_matching: bool,
+) -> (IndexBuilder, TempDir, Arc<RwLock<DB>>) {
+    let temp_dir = Builder::new().prefix("test_dir").tempdir().unwrap();
+    #[cfg(feature = "rocksdb")]
+    let db = open_db_with_existing_cf(&temp_dir.path().join("test_db")).unwrap();
+    #[cfg(not(feature = "rocksdb"))]
+    let db = ();
+
+    let config = TextIndexParams {
+        phrase_matching: Some(phrase_matching),
+        ..TextIndexParams::default()
+    };
+
+    let mut builder = match index_type {
+        #[cfg(feature = "rocksdb")]
+        IndexType::Mutable => IndexBuilder::Mutable(FullTextIndex::builder_rocksdb(
+            db.clone(),
+            config,
+            FIELD_NAME,
+        )),
+        IndexType::MutableGridstore => IndexBuilder::MutableGridstore(
+            FullTextIndex::builder_gridstore(temp_dir.path().to_path_buf(), config),
+        ),
+        #[cfg(feature = "rocksdb")]
+        IndexType::Immutable => IndexBuilder::Immutable(FullTextIndex::builder_rocksdb(
+            db.clone(),
+            config,
+            FIELD_NAME,
+        )),
+        IndexType::Mmap => IndexBuilder::Mmap(FullTextIndex::builder_mmap(
+            temp_dir.path().to_path_buf(),
+            config,
+            true,
+        )),
+        IndexType::RamMmap => IndexBuilder::RamMmap(FullTextIndex::builder_mmap(
+            temp_dir.path().to_path_buf(),
+            config,
+            false,
+        )),
+    };
+    match &mut builder {
+        #[cfg(feature = "rocksdb")]
+        IndexBuilder::Mutable(builder) => builder.init().unwrap(),
+        IndexBuilder::MutableGridstore(builder) => builder.init().unwrap(),
+        #[cfg(feature = "rocksdb")]
+        IndexBuilder::Immutable(builder) => builder.init().unwrap(),
+        IndexBuilder::Mmap(builder) => builder.init().unwrap(),
+        IndexBuilder::RamMmap(builder) => builder.init().unwrap(),
+    }
+    (builder, temp_dir, db)
+}
+
+fn build_random_index(
+    num_points: usize,
+    num_keywords: usize,
+    keyword_len: usize,
+    index_type: IndexType,
+    phrase_matching: bool,
+    deleted: bool,
+) -> (FullTextIndex, TempDir, Arc<RwLock<DB>>) {
+    let mut rnd = StdRng::seed_from_u64(42);
+    let (mut builder, temp_dir, db) = create_builder(index_type, phrase_matching);
+
+    for idx in 0..num_points {
+        let keywords = random_full_text_payload(
+            &mut rnd,
+            num_keywords..=num_keywords,
+            keyword_len..=keyword_len,
+        );
+        let array_payload = Value::Array(keywords);
+        builder
+            .add_point(
+                idx as PointOffsetType,
+                &[&array_payload],
+                &HardwareCounterCell::new(),
+            )
+            .unwrap();
+    }
+
+    let mut index = builder.finalize().unwrap();
+    assert_eq!(index.points_count(), num_points);
+
+    // Delete some points before loading into a different format
+    if deleted {
+        index.remove_point(20).unwrap();
+        index.remove_point(21).unwrap();
+        index.remove_point(22).unwrap();
+        index.remove_point(200).unwrap();
+        index.remove_point(250).unwrap();
+    }
+
+    (index, temp_dir, db)
+}
+
+/// Tries to parse a query. If there is an unknown id to a token, returns `None`
+pub fn to_parsed_query(
+    query: &[String],
+    token_to_id: impl Fn(&str) -> Option<TokenId>,
+) -> Option<ParsedQuery> {
+    let tokens = query
+        .iter()
+        .map(|token| token_to_id(token.as_str()))
+        .collect::<Option<TokenSet>>()?;
+    Some(ParsedQuery::Tokens(tokens))
+}
+
+pub fn parse_query(query: &[String], index: &FullTextIndex) -> ParsedQuery {
+    let hw_counter = HardwareCounterCell::disposable();
+    match index {
+        FullTextIndex::Mutable(index) => {
+            let token_to_id = |token: &str| index.inverted_index.get_token_id(token, &hw_counter);
+            to_parsed_query(query, token_to_id).unwrap()
+        }
+        FullTextIndex::Immutable(index) => {
+            let token_to_id = |token: &str| index.inverted_index.get_token_id(token, &hw_counter);
+            to_parsed_query(query, token_to_id).unwrap()
+        }
+        FullTextIndex::Mmap(index) => {
+            let token_to_id = |token: &str| index.inverted_index.get_token_id(token, &hw_counter);
+            to_parsed_query(query, token_to_id).unwrap()
+        }
+    }
+}
+
+#[rstest]
+fn test_congruence(
+    #[values(false, true)] deleted: bool,
+    #[values(false, true)] phrase_matching: bool,
+) {
+    const POINT_COUNT: usize = 500;
+    const KEYWORD_COUNT: usize = 5;
+    const KEYWORD_LEN: usize = 2;
+
+    let hw_counter = HardwareCounterCell::disposable();
+
+    let (mut indices, _data): (Vec<_>, Vec<_>) = TYPES
+        .iter()
+        .copied()
+        .map(|index_type| {
+            let (index, temp_dir, db) = build_random_index(
+                POINT_COUNT,
+                KEYWORD_COUNT,
+                KEYWORD_LEN,
+                index_type,
+                phrase_matching,
+                deleted,
+            );
+            ((index, index_type), (temp_dir, db))
+        })
+        .unzip();
+
+    // Delete some points after loading
+    if deleted {
+        for (index, _type) in indices.iter_mut() {
+            index.remove_point(10).unwrap();
+            index.remove_point(11).unwrap();
+            index.remove_point(12).unwrap();
+            index.remove_point(100).unwrap();
+            index.remove_point(150).unwrap();
+        }
+    }
+
+    // Grab 10 keywords to use for querying
+    let (FullTextIndex::Mutable(index), _) = &indices[0] else {
+        panic!("Expects mutable full text index as first");
+    };
+    let mut keywords = index
+        .inverted_index
+        .vocab
+        .keys()
+        .cloned()
+        .collect::<Vec<_>>();
+    keywords.sort_unstable();
+    keywords.truncate(10);
+
+    for i in 1..indices.len() {
+        use common::types::PointOffsetType;
+
+        let ((index_a, type_a), (index_b, type_b)) = (&indices[0], &indices[i]);
+        eprintln!("Testing index type {type_a:?} vs {type_b:?}");
+
+        assert_eq!(index_a.points_count(), index_b.points_count());
+        for point_id in 0..POINT_COUNT as PointOffsetType {
+            assert_eq!(
+                index_a.values_count(point_id),
+                index_b.values_count(point_id),
+            );
+            assert_eq!(
+                index_a.values_is_empty(point_id),
+                index_b.values_is_empty(point_id),
+            );
+        }
+
+        assert_eq!(
+            index_a.get_token("doesnotexist", &hw_counter),
+            index_b.get_token("doesnotexist", &hw_counter),
+        );
+        assert!(
+            index_a.get_token(&keywords[0], &hw_counter).is_some()
+                == index_b.get_token(&keywords[0], &hw_counter).is_some(),
+        );
+
+        for query_range in [0..1, 2..4, 5..9, 0..10] {
+            let keywords = &keywords[query_range];
+            let parsed_query_a = parse_query(keywords, index_a);
+            let parsed_query_b = parse_query(keywords, index_b);
+
+            // Mutable index behaves different versus the others on point deletion
+            // Mutable index updates postings, the others do not. Cardinality estimations are
+            // not expected to match because of it.
+            if !deleted {
+                use crate::types::{FieldCondition, ValuesCount};
+
+                let field_condition = FieldCondition::new_values_count(
+                    JsonPath::new(FIELD_NAME),
+                    ValuesCount::from(0..10),
+                );
+                let cardinality_a = index_a.estimate_query_cardinality(
+                    &parsed_query_a,
+                    &field_condition,
+                    &hw_counter,
+                );
+                let cardinality_b = index_b.estimate_query_cardinality(
+                    &parsed_query_b,
+                    &field_condition,
+                    &hw_counter,
+                );
+                assert_eq!(cardinality_a, cardinality_b);
+            }
+
+            for point_id in 0..POINT_COUNT as PointOffsetType {
+                assert_eq!(
+                    index_a.check_match(&parsed_query_a, point_id, &hw_counter),
+                    index_b.check_match(&parsed_query_b, point_id, &hw_counter),
+                );
+            }
+
+            assert_eq!(
+                index_a
+                    .filter_query(parsed_query_a, &hw_counter)
+                    .collect::<HashSet<_>>(),
+                index_b
+                    .filter_query(parsed_query_b, &hw_counter)
+                    .collect::<HashSet<_>>(),
+            );
+        }
+
+        if !deleted {
+            for threshold in 1..=10 {
+                assert_eq!(
+                    index_a
+                        .payload_blocks(threshold, JsonPath::new(FIELD_NAME))
+                        .count(),
+                    index_b
+                        .payload_blocks(threshold, JsonPath::new(FIELD_NAME))
+                        .count(),
+                );
+            }
+        }
+    }
+}
